apps/task-manager/backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
psycopg2-binary>=2.9
gunicorn>=21.0
